requirements.txt
geographiclib==1.49
geopy==1.17.0
numpy==1.15.4
pandas==0.23.4
pkg-resources==0.0.0
pycountry==18.5.26
python-dateutil==2.7.5
python-slugify==1.2.6
pytz==2018.7
six==1.11.0
Unidecode==1.0.23
